requirements.txt
Flask==3.1.1
python-dotenv==1.1.1
psycopg2-binary==2.9.10
pillow==11.3.0 